rds-comparison/benchmark/requirements.txt
psycopg2-binary>=2.9
pymysql>=1.1
boto3>=1.34
matplotlib>=3.8
